+++ noodles-sam/src/alignment_reader.rs
@@ -1,7 +1,5 @@
 use std::io;
 
-use noodles_fasta as fasta;
-
 use super::{alignment::Record, Header};
 
 /// An alignment format reader.
@@ -12,7 +10,6 @@ pub trait AlignmentReader<R> {
     /// Returns an iterator over records.
     fn alignment_records<'a>(
         &'a mut self,
-        reference_sequence_repository: &'a fasta::Repository,
         header: &'a Header,
     ) -> Box<dyn Iterator<Item = io::Result<Record>> + 'a>;
 }