+++ noodles-sam/src/reader.rs
@@ -13,7 +13,6 @@ use std::io::{self, BufRead, Read, Seek};
 use noodles_bgzf as bgzf;
 use noodles_core::Region;
 use noodles_csi as csi;
-use noodles_fasta as fasta;
 
 use super::{alignment::Record, header::ReferenceSequences, lazy, AlignmentReader, Header};
 
@@ -334,7 +333,6 @@ where
 
     fn alignment_records<'a>(
         &'a mut self,
-        _: &'a fasta::Repository,
         header: &'a Header,
     ) -> Box<dyn Iterator<Item = io::Result<Record>> + 'a> {
         Box::new(self.records(header))